refactor: adjust utils and types in testing folder
