+++ src/__tests__/types.ts
@@ -1,11 +1,3 @@
-export type PostLayout = 'classic' | 'image';
-
-export interface Post {
-  id: number;
-  title: string;
-  layout: PostLayout;
-}
-
 export type Func = (...args: any[]) => any;
 
 export interface PostModelControl {
