+++ src/__tests__/useAccessor-normal-dedupeInterval.test.tsx
@@ -1,10 +1,10 @@
 import { useAccessor } from '../lib';
 import { act, render, screen } from '@testing-library/react';
-import { createPostModel, createPostModelControl, sleep } from './utils';
+import { createPostModel, createControl, sleep } from './utils';
 
 describe('useAccessor-normal dedupeInterval', () => {
   test('should sync model with the data from the latest request', async () => {
-    const control = createPostModelControl({ sleepTime: 100 });
+    const control = createControl({ sleepTime: 100 });
     const { postAdapter, getPostById } = createPostModel(control);
     const accessor = getPostById(0);
     function Page() {
